Handle when a non-URL plain text attachment is included
Previously we always looked at the first attachment. But Brave (at least when tapping-and-holding a link) includes the name of it as the first attachment, and the URL as the second. So we do two passes: first we loop through the attachments and look for a URL. If one is not found, then we loop through again and look for plain text. Because NSItemProvider.loadItem() is asynchronous, it would be more complex to *test* each plain text item to see if it can be converted to a URL, so we won't do that until we run into a case where it's needed. Instead, we just go with the first one we find.